+++ src/main/java/com/coastee/server/auth/AdminOnly.java
@@ -1,0 +1,12 @@
+package com.coastee.server.auth;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.Target;
+
+import static java.lang.annotation.ElementType.METHOD;
+import static java.lang.annotation.RetentionPolicy.RUNTIME;
+
+@Target(METHOD)
+@Retention(RUNTIME)
+public @interface AdminOnly {
+}
