+++ src/main/java/com/coastee/server/auth/Auth.java
@@ -1,0 +1,12 @@
+package com.coastee.server.auth;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.Target;
+
+import static java.lang.annotation.ElementType.PARAMETER;
+import static java.lang.annotation.RetentionPolicy.RUNTIME;
+
+@Target(PARAMETER)
+@Retention(RUNTIME)
+public @interface Auth {
+}
