+++ src/main/java/com/coastee/server/global/config/JwtAuthorizationFilter.java
@@ -2,7 +2,7 @@ package com.coastee.server.global.config;
 
 import com.coastee.server.global.apipayload.code.ErrorReasonDTO;
 import com.coastee.server.global.apipayload.exception.GeneralException;
-import com.coastee.server.jwt.util.JwtProvider;
+import com.coastee.server.login.infrastructure.JwtProvider;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.jsonwebtoken.Claims;
 import jakarta.servlet.FilterChain;
