--- src/main/java/com/coastee/server/login/auth/AdminAuth.java
@@ -1,12 +1,0 @@
-package com.coastee.server.login.auth;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.Target;
-
-import static java.lang.annotation.ElementType.PARAMETER;
-import static java.lang.annotation.RetentionPolicy.RUNTIME;
-
-@Target(PARAMETER)
-@Retention(RUNTIME)
-public @interface AdminAuth {
-}
