--- src/main/java/com/coastee/server/login/auth/domain/Authority.java
@@ -1,27 +1,0 @@
-package com.coastee.server.login.auth.domain;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-import java.util.Arrays;
-
-@Getter
-@RequiredArgsConstructor
-public enum Authority {
-    USER("USER"),
-    ADMIN("ADMIN");
-
-    private final String code;
-
-    @Override
-    public String toString() {
-        return this.code;
-    }
-
-    public static Authority of(final String code) {
-        return Arrays.stream(Authority.values())
-                .filter(r -> r.getCode().equals(code.toUpperCase()))
-                .findAny()
-                .orElse(null);
-    }
-}
