+++ src/main/java/com/coastee/server/login/domain/OAuthUserInfo.java
@@ -1,0 +1,14 @@
+package com.coastee.server.login.domain;
+
+import com.coastee.server.user.domain.SocialType;
+
+public interface OAuthUserInfo {
+
+    String getSocialId();
+
+    String getName();
+
+    String getEmail();
+
+    SocialType getSocialType();
+}
