+++ src/main/java/com/coastee/server/login/domain/SocialTokens.java
@@ -1,0 +1,8 @@
+package com.coastee.server.login.domain;
+
+public interface SocialTokens {
+
+    String getAccessToken();
+
+    String getRefreshToken();
+}
