--- src/main/java/com/coastee/server/login/jwt/domain/AuthTokens.java
@@ -1,21 +1,0 @@
-package com.coastee.server.login.jwt.domain;
-
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class AuthTokens {
-    private String accessToken;
-    private String refreshToken;
-
-    public static AuthTokens of(
-            final String accessToken,
-            final String refreshToken
-    ) {
-        return new AuthTokens(accessToken, refreshToken);
-    }
-}
