--- src/main/java/com/coastee/server/oauth/domain/NaverSocialTokens.java
@@ -1,31 +1,0 @@
-package com.coastee.server.oauth.domain;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class NaverSocialTokens implements SocialTokens {
-
-    @JsonProperty("access_token")
-    private String accessToken;
-
-    @JsonProperty("refresh_token")
-    private String refreshToken;
-
-    @JsonProperty("token_type")
-    private String tokenType;
-
-    @JsonProperty("expires_in")
-    private String expiresIn;
-
-    @JsonProperty("error")
-    private String error;
-
-    @JsonProperty("error_description")
-    private String errorDescription;
-}
