--- src/main/java/com/coastee/server/oauth/domain/OAuthLoginParams.java
@@ -1,12 +1,0 @@
-package com.coastee.server.oauth.domain;
-
-import com.coastee.server.user.domain.SocialType;
-
-public interface OAuthLoginParams {
-
-    SocialType socialType();
-
-    void updateClientId(final String clientId);
-
-    void updateClientSecret(final String clientSecret);
-}
