--- src/main/java/com/coastee/server/oauth/domain/SocialTokens.java
@@ -1,8 +1,0 @@
-package com.coastee.server.oauth.domain;
-
-public interface SocialTokens {
-
-    String getAccessToken();
-
-    String getRefreshToken();
-}
