--- src/main/java/com/coastee/server/oauth/dto/info/OAuthInfoResponse.java
@@ -1,14 +1,0 @@
-package com.coastee.server.oauth.dto.info;
-
-import com.coastee.server.user.domain.SocialType;
-
-public interface OAuthInfoResponse {
-
-    String getSocialId();
-
-    String getName();
-
-    String getEmail();
-
-    SocialType getSocialType();
-}
